requirements.txt
streamlit
tensorflow
numpy
pandas
pillow
matplotlib
seaborn
scikit-learn
lime
shap
